+++ webapp/user/views.py
@@ -5,10 +5,10 @@ from webapp.db import db
 from webapp.user.forms import LoginForm, RegistrationForm, SmsAuthForm, UpdateDataProfileUserForm
 from webapp.user.models import User
 from webapp.user.enums import EmailEventsForUser, SmsEventsForUser, UserRole
-from webapp.user.tasks import send_email, send_sms
-from webapp.services.service_send_sms import delete_symbols_from_phone_number, generate_six_digits_code
+from webapp.services.service_send_sms import delete_symbols_from_phone_number, generate_six_digits_code, send_sms
 from webapp.services.service_redirect_utils import redirect_back
 from webapp.services.service_cart import save_products_into_db_from_session_cart
+from webapp.services.service_send_email import send_email
 
 blueprint = Blueprint('user', __name__, url_prefix='/users')
 
@@ -84,8 +84,9 @@ def process_reg():
         new_user.set_password(form.password.data)
         db.session.add(new_user)
         db.session.commit()
-        flash('Вы успешно зарегистрировались')
-        send_email.delay(EmailEventsForUser.hello_letter, new_user)
+        flash('Вы успешно зарегистрировались и авторизовались')
+        send_email(EmailEventsForUser.hello_letter, new_user)
+        login_user(new_user)
         return redirect(url_for('marketplace.index'))
 
     else:
@@ -168,7 +169,7 @@ def process_auth_sms():
             db.session.commit()
 
         code_for_sms = generate_six_digits_code()
-        send_sms_with_code = send_sms.delay(SmsEventsForUser.send_auth_sms, user, generated_code=code_for_sms)
+        send_sms_with_code = send_sms(SmsEventsForUser.send_auth_sms, user, generated_code=code_for_sms)
 
         if send_sms_with_code:
             session['sms_code'] = code_for_sms
